Remove old source files. Related to #761.
