+++ gqcp/include/ONVBasis/CMakeLists.txt
@@ -1,7 +1,5 @@
 target_sources(gqcp
     PRIVATE
-        BaseONVBasis.hpp
-        ONVManipulator.hpp
         ONVPath.hpp
         SeniorityZeroONVBasis.hpp
         SpinResolvedONV.hpp
