+++ gqcp/include/QCMethod/CI/CI.hpp
@@ -21,7 +21,6 @@
 #include "Mathematical/Algorithm/Algorithm.hpp"
 #include "Mathematical/Algorithm/IterativeAlgorithm.hpp"
 #include "Mathematical/Optimization/Eigenproblem/EigenproblemEnvironment.hpp"
-#include "ONVBasis/BaseONVBasis.hpp"
 #include "QCMethod/QCStructure.hpp"
 #include "QCModel/CI/LinearExpansion.hpp"
