--- gqcp/include/QCMethod/CI/HamiltonianBuilder/CMakeLists.txt
@@ -1,8 +1,0 @@
-target_sources(gqcp
-    PRIVATE
-        DOCI.hpp
-        FCI.hpp
-        HamiltonianBuilder.hpp
-        Hubbard.hpp
-        SelectedCI.hpp
-)
