--- gqcp/src/QCMethod/CI/HamiltonianBuilder/CMakeLists.txt
@@ -1,7 +1,0 @@
-target_sources(gqcp
-    PRIVATE
-        DOCI.cpp
-        FCI.cpp
-        Hubbard.cpp
-        SelectedCI.cpp
-)
