+++ gqcp/src/QCMethod/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_subdirectory(Applications)
-add_subdirectory(CI)
 add_subdirectory(Geminals)
 add_subdirectory(OrbitalOptimization)
 add_subdirectory(RMP2)
